requirements.txt
tk
pillow
opencv-python
numpy
pandas
spellchecker
tensorflow
keras
